feature(summary) #446: some refactor about summary. 1. Add min_seq_no in `MemCache`, if it's smaller than Version::last_seq, we can safely remove it. 2. When call Summary::write_summary(), dispatch `VersionEdit`s for each TsFamily (maps TsFamilyId to `VersionEdit`s), then produce new `Version`s and `SuperVersion`s for them. 3. `Version` now can use a series of `VersionEdit` to generate a new `Version`. 4. TsFamily now has an `Arc<Version>` and an `Arc<dyn Picker>`, and some other refactors for pickers. 5. Reduce arguments for VersionEdit::add_file(). 6. Implement the Default trait for `CompactMeta`. 7. Exchange the two arguments for TimeRange::new().

+++ tskv/src/tsm/reader.rs
@@ -385,7 +385,6 @@ impl BlockMetaIterator {
         while pos < sli.len() {
             self.block_meta_limit += 1;
             if max_ts < decode_be_i64(&sli[pos + 8..pos + 16]) {
-                // Last data block in time range
                 return;
             } else {
                 pos += BLOCK_META_SIZE;
@@ -580,7 +579,7 @@ pub fn decode_data_block(
     }
 
     // let crc_ts = &self.buf[..4];
-    let mut ts = Vec::with_capacity(MAX_BLOCK_VALUES);
+    let mut ts = Vec::with_capacity(MAX_BLOCK_VALUES as usize);
     timestamp::decode(&buf[4..val_off as usize], &mut ts).context(DecodeSnafu)?;
     // let crc_data = &self.buf[(val_offset - offset) as usize..4];
     let data = &buf[(val_off + 4) as usize..];
